+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.0.2
 * Fix empty reply message due to incorrect random index
+* Retrieve Dialogflow supported languages from Convo specification #2
 
 ### 0.0.1
 * Initial version